Revert "Next.js 13 - App directory"
