--- app/about/head.tsx
@@ -1,19 +1,0 @@
-import { HeadMeta } from "../../components/HeadMeta";
-import DefaultTags from "../DefaultTags";
-
-const meta = {
-	title: "Milovan Gudelj - About me",
-	description:
-		"I design and develop engaging websites and delightful digital experiences.",
-	url: "https://milovangudelj.com/about",
-	image: "https://milovangudelj.com/images/og-image.png",
-};
-
-export default function RootHead() {
-	return (
-		<>
-			<DefaultTags />
-			<HeadMeta metadata={meta} />
-		</>
-	);
-}
