+++ components/NowPlaying/NowPlaying.tsx
@@ -1,5 +1,3 @@
-"use client";
-
 import { useState } from "react";
 import Image from "next/image";
 import useSWR from "swr";
@@ -97,5 +95,3 @@ export const NowPlaying = () => {
 		</div>
 	);
 };
-
-export default NowPlaying;
