+++ components/WrappedList/WrappedList.tsx
@@ -238,5 +238,3 @@ const YoungAccount = ({
 		</div>
 	);
 };
-
-export default WrappedList;
