--- lib/sessionProvider.tsx
@@ -1,5 +1,0 @@
-"use client";
-
-import { SessionProvider as Provider } from "next-auth/react";
-
-export const SessionProvider = Provider;
