+++ package.json
@@ -14,7 +14,6 @@
 		"@next/font": "^13.1.6",
 		"@vercel/analytics": "^0.1.8",
 		"cva": "npm:class-variance-authority",
-		"framer-motion": "^9.0.2",
 		"graphql": "16.6.0",
 		"graphql-request": "5.1.0",
 		"html-to-image": "^1.11.1",
@@ -27,7 +26,7 @@
 		"react": "^18.2.0",
 		"react-dom": "^18.2.0",
 		"react-hook-form": "^7.40.0",
-		"sharp": "^0.31.3",
+		"sharp": "^0.31.2",
 		"spotify-web-api-node": "^5.0.2",
 		"swr": "^2.0.0",
 		"tailwind-merge": "^1.8.0"
