+++ utils/useIsomorphicLayoutEffect.tsx
@@ -1,6 +1,4 @@
-"use client";
-
-import { useLayoutEffect } from "react";
+import * as React from "react";
 
 /**
  * On the server, React emits a warning when calling `useLayoutEffect`.
@@ -10,6 +8,6 @@ import { useLayoutEffect } from "react";
  * See: https://reactjs.org/docs/hooks-reference.html#uselayouteffect
  */
 const useIsomorphicLayoutEffect =
-	typeof window !== "undefined" ? useLayoutEffect : () => {};
+	typeof window !== "undefined" ? React.useLayoutEffect : () => {};
 
 export { useIsomorphicLayoutEffect };
